Moved json serialization out of passive classes
