+++ src/main/java/customizables/passives/OnBeHitPassive.java
@@ -3,11 +3,8 @@ package customizables.passives;
 import actions.OnHitEvent;
 import actions.OnHitListener;
 import entities.AbstractPlayer;
-import javax.json.JsonObject;
-import serialization.JsonSerialable;
-import customizables.CustomizableJsonUtil;
 
-public class OnBeHitPassive extends AbstractPassive implements JsonSerialable, OnHitListener{
+public class OnBeHitPassive extends AbstractPassive implements OnHitListener{
 	/**
 	 * Triggers once the user's hitbox intercepts 
 	 * that of an enemy projectile
@@ -22,20 +19,6 @@ public class OnBeHitPassive extends AbstractPassive implements JsonSerialable, O
 		copyInflictTo(copy);
 		return copy;
 	}
-    
-    @Override
-    public JsonObject serializeJson(){
-        return super.serializeJson();
-    }
-    
-    public static OnBeHitPassive deserializeJson(JsonObject obj){
-        OnBeHitPassive obh = new OnBeHitPassive(
-            CustomizableJsonUtil.getNameFrom(obj),
-            getTargetsUserFrom(obj)
-        );
-        obh.setInflict(CustomizableJsonUtil.getStatusTableFrom(obj));
-        return obh;
-    }
     
     @Override
     public void init(){
